+++ console2/src/api/service/console/index.ts
@@ -91,6 +91,7 @@ export interface RepositoryTestRequest {
     branch?: string;
     commitId?: string;
     path?: string;
+    withSecret?: boolean;
     secretId?: string;
     secretName?: string;
 }